o/devicestate,o/install: use overlord/install to check for encryption support
