+++ overlord/devicestate/devicestate_install_mode_test.go
@@ -42,7 +42,6 @@ import (
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/gadget"
 	"github.com/snapcore/snapd/gadget/install"
-	"github.com/snapcore/snapd/gadget/quantity"
 	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/osutil"
 	"github.com/snapcore/snapd/overlord/assertstate/assertstatetest"
@@ -51,6 +50,7 @@ import (
 	"github.com/snapcore/snapd/overlord/devicestate/devicestatetest"
 	"github.com/snapcore/snapd/overlord/hookstate"
 	"github.com/snapcore/snapd/overlord/hookstate/ctlcmd"
+	installLogic "github.com/snapcore/snapd/overlord/install"
 	"github.com/snapcore/snapd/overlord/restart"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/snapstate/snapstatetest"
@@ -108,7 +108,7 @@ func (s *deviceMgrInstallModeSuite) SetUpTest(c *C) {
 	})
 	s.AddCleanup(restore)
 
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
 		c.Check(tpmMode, Equals, secboot.TPMProvisionFull)
 		return fmt.Errorf("TPM not available")
 	})
@@ -267,7 +267,7 @@ func (s *deviceMgrInstallModeSuite) doRunChangeTestWithEncryption(c *C, grade st
 	})
 	defer restore()
 
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
 		c.Check(tpmMode, Equals, secboot.TPMProvisionFull)
 		if tc.tpm {
 			return nil
@@ -1483,7 +1483,7 @@ func (s *deviceMgrInstallModeSuite) TestInstallBootloaderVarSetFails(c *C) {
 	})
 	defer restore()
 
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return fmt.Errorf("no encrypted soup for you") })
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return fmt.Errorf("no encrypted soup for you") })
 	defer restore()
 
 	err := ioutil.WriteFile(filepath.Join(dirs.GlobalRootDir, "/var/lib/snapd/modeenv"),
@@ -1512,7 +1512,7 @@ func (s *deviceMgrInstallModeSuite) testInstallEncryptionValidityChecks(c *C, er
 	restore := release.MockOnClassic(false)
 	defer restore()
 
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return nil })
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return nil })
 	defer restore()
 
 	err := ioutil.WriteFile(filepath.Join(dirs.GlobalRootDir, "/var/lib/snapd/modeenv"),
@@ -1847,7 +1847,7 @@ func (s *deviceMgrInstallModeSuite) TestInstallWithEncryptionValidatesGadgetErr(
 	defer restore()
 
 	// pretend we have a TPM
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
 		c.Check(tpmMode, Equals, secboot.TPMProvisionFull)
 		return nil
 	})
@@ -1879,7 +1879,7 @@ func (s *deviceMgrInstallModeSuite) TestInstallWithEncryptionValidatesGadgetWarn
 	defer restore()
 
 	// pretend we have a TPM
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return nil })
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return nil })
 	defer restore()
 
 	s.testInstallGadgetNoSave(c, "dangerous")
@@ -1903,7 +1903,7 @@ func (s *deviceMgrInstallModeSuite) TestInstallWithoutEncryptionValidatesGadgetW
 	defer restore()
 
 	// pretend we have a TPM
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return fmt.Errorf("TPM2 not available") })
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return fmt.Errorf("TPM2 not available") })
 	defer restore()
 
 	s.testInstallGadgetNoSave(c, "dangerous")
@@ -1979,7 +1979,7 @@ func (s *deviceMgrInstallModeSuite) TestInstallCheckEncrypted(c *C) {
 		} else {
 			makeInstalledMockKernelSnap(c, st, kernelYamlNoFdeSetup)
 		}
-		restore := devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error {
+		restore := installLogic.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error {
 			if tc.hasTPM {
 				return nil
 			}
@@ -1995,238 +1995,6 @@ func (s *deviceMgrInstallModeSuite) TestInstallCheckEncrypted(c *C) {
 		}
 		logbuf.Reset()
 	}
-}
-
-func (s *deviceMgrInstallModeSuite) TestInstallCheckEncryptedStorageSafety(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	s.makeMockInstalledPcKernelAndGadget(c, "", "")
-
-	restore := devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return nil })
-	defer restore()
-
-	var testCases = []struct {
-		grade, storageSafety string
-
-		expectedEncryption bool
-	}{
-		// we don't test unset here because the assertion assembly
-		// will ensure it has a default
-		{"dangerous", "prefer-unencrypted", false},
-		{"dangerous", "prefer-encrypted", true},
-		{"dangerous", "encrypted", true},
-		{"signed", "prefer-unencrypted", false},
-		{"signed", "prefer-encrypted", true},
-		{"signed", "encrypted", true},
-		// secured+prefer-{,un}encrypted is an error at the
-		// assertion level already so cannot be tested here
-		{"secured", "encrypted", true},
-	}
-	for _, tc := range testCases {
-		mockModel := s.makeModelAssertionInState(c, "my-brand", "my-model", map[string]interface{}{
-			"display-name":   "my model",
-			"architecture":   "amd64",
-			"base":           "core20",
-			"grade":          tc.grade,
-			"storage-safety": tc.storageSafety,
-			"snaps": []interface{}{
-				map[string]interface{}{
-					"name":            "pc-kernel",
-					"id":              pcKernelSnapID,
-					"type":            "kernel",
-					"default-channel": "20",
-				},
-				map[string]interface{}{
-					"name":            "pc",
-					"id":              pcSnapID,
-					"type":            "gadget",
-					"default-channel": "20",
-				}},
-		})
-		deviceCtx := &snapstatetest.TrivialDeviceContext{DeviceModel: mockModel}
-
-		encryptionType, err := devicestate.DeviceManagerCheckEncryption(s.mgr, s.state, deviceCtx, secboot.TPMProvisionFull)
-		c.Assert(err, IsNil)
-		encrypt := (encryptionType != secboot.EncryptionTypeNone)
-		c.Check(encrypt, Equals, tc.expectedEncryption, Commentf("%v", tc))
-	}
-}
-
-func (s *deviceMgrInstallModeSuite) TestInstallCheckEncryptedErrors(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	s.makeMockInstalledPcKernelAndGadget(c, "", "")
-
-	restore := devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return fmt.Errorf("tpm says no") })
-	defer restore()
-
-	var testCases = []struct {
-		grade, storageSafety string
-
-		expectedErr string
-	}{
-		// we don't test unset here because the assertion assembly
-		// will ensure it has a default
-		{
-			"dangerous", "encrypted",
-			"cannot encrypt device storage as mandated by encrypted storage-safety model option: tpm says no",
-		}, {
-			"signed", "encrypted",
-			"cannot encrypt device storage as mandated by encrypted storage-safety model option: tpm says no",
-		}, {
-			"secured", "",
-			"cannot encrypt device storage as mandated by model grade secured: tpm says no",
-		}, {
-			"secured", "encrypted",
-			"cannot encrypt device storage as mandated by model grade secured: tpm says no",
-		},
-	}
-	for _, tc := range testCases {
-		mockModel := s.makeModelAssertionInState(c, "my-brand", "my-model", map[string]interface{}{
-			"display-name":   "my model",
-			"architecture":   "amd64",
-			"base":           "core20",
-			"grade":          tc.grade,
-			"storage-safety": tc.storageSafety,
-			"snaps": []interface{}{
-				map[string]interface{}{
-					"name":            "pc-kernel",
-					"id":              pcKernelSnapID,
-					"type":            "kernel",
-					"default-channel": "20",
-				},
-				map[string]interface{}{
-					"name":            "pc",
-					"id":              pcSnapID,
-					"type":            "gadget",
-					"default-channel": "20",
-				}},
-		})
-		deviceCtx := &snapstatetest.TrivialDeviceContext{DeviceModel: mockModel}
-		_, err := devicestate.DeviceManagerCheckEncryption(s.mgr, s.state, deviceCtx, secboot.TPMProvisionFull)
-		c.Check(err, ErrorMatches, tc.expectedErr, Commentf("%s %s", tc.grade, tc.storageSafety))
-	}
-}
-
-func (s *deviceMgrInstallModeSuite) TestInstallCheckEncryptedFDEHook(c *C) {
-	st := s.state
-	st.Lock()
-	defer st.Unlock()
-
-	s.makeModelAssertionInState(c, "canonical", "pc", map[string]interface{}{
-		"architecture": "amd64",
-		"kernel":       "pc-kernel",
-		"gadget":       "pc",
-	})
-	devicestatetest.SetDevice(s.state, &auth.DeviceState{
-		Brand: "canonical",
-		Model: "pc",
-	})
-	makeInstalledMockKernelSnap(c, st, kernelYamlWithFdeSetup)
-
-	for _, tc := range []struct {
-		hookOutput  string
-		expectedErr string
-
-		encryptionType secboot.EncryptionType
-	}{
-		// invalid json
-		{"xxx", `cannot parse hook output "xxx": invalid character 'x' looking for beginning of value`, secboot.EncryptionTypeNone},
-		// no output is invalid
-		{"", `cannot parse hook output "": unexpected end of JSON input`, secboot.EncryptionTypeNone},
-		// specific error
-		{`{"error":"failed"}`, `cannot use hook: it returned error: failed`, secboot.EncryptionTypeNone},
-		{`{}`, `cannot use hook: neither "features" nor "error" returned`, secboot.EncryptionTypeNone},
-		// valid
-		{`{"features":[]}`, "", secboot.EncryptionTypeLUKS},
-		{`{"features":["a"]}`, "", secboot.EncryptionTypeLUKS},
-		{`{"features":["a","b"]}`, "", secboot.EncryptionTypeLUKS},
-		// features must be list of strings
-		{`{"features":[1]}`, `cannot parse hook output ".*": json: cannot unmarshal number into Go struct.*`, secboot.EncryptionTypeNone},
-		{`{"features":1}`, `cannot parse hook output ".*": json: cannot unmarshal number into Go struct.*`, secboot.EncryptionTypeNone},
-		{`{"features":"1"}`, `cannot parse hook output ".*": json: cannot unmarshal string into Go struct.*`, secboot.EncryptionTypeNone},
-		// valid and uses ice
-		{`{"features":["a","inline-crypto-engine","b"]}`, "", secboot.EncryptionTypeLUKSWithICE},
-	} {
-		hookInvoke := func(ctx *hookstate.Context, tomb *tomb.Tomb) ([]byte, error) {
-			ctx.Lock()
-			defer ctx.Unlock()
-			ctx.Set("fde-setup-result", []byte(tc.hookOutput))
-			return nil, nil
-		}
-		rhk := hookstate.MockRunHook(hookInvoke)
-		defer rhk()
-
-		et, err := devicestate.DeviceManagerCheckFDEFeatures(s.mgr, st)
-		if tc.expectedErr != "" {
-			c.Check(err, ErrorMatches, tc.expectedErr, Commentf("%v", tc))
-		} else {
-			c.Check(err, IsNil, Commentf("%v", tc))
-			c.Check(et, Equals, tc.encryptionType, Commentf("%v", tc))
-		}
-	}
-}
-
-var checkEncryptionModelHeaders = map[string]interface{}{
-	"display-name": "my model",
-	"architecture": "amd64",
-	"base":         "core20",
-	"grade":        "dangerous",
-	"snaps": []interface{}{
-		map[string]interface{}{
-			"name":            "pc-kernel",
-			"id":              pcKernelSnapID,
-			"type":            "kernel",
-			"default-channel": "20",
-		},
-		map[string]interface{}{
-			"name":            "pc",
-			"id":              pcSnapID,
-			"type":            "gadget",
-			"default-channel": "20",
-		}},
-}
-
-func (s *deviceMgrInstallModeSuite) TestInstallCheckEncryptedErrorsLogsTPM(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	s.makeMockInstalledPcKernelAndGadget(c, "", "")
-
-	restore := devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error {
-		return fmt.Errorf("tpm says no")
-	})
-	defer restore()
-
-	logbuf, restore := logger.MockLogger()
-	defer restore()
-
-	mockModel := s.makeModelAssertionInState(c, "my-brand", "my-model", checkEncryptionModelHeaders)
-	deviceCtx := &snapstatetest.TrivialDeviceContext{DeviceModel: mockModel}
-	_, err := devicestate.DeviceManagerCheckEncryption(s.mgr, s.state, deviceCtx, secboot.TPMProvisionFull)
-	c.Check(err, IsNil)
-	c.Check(logbuf.String(), Matches, "(?s).*: not encrypting device storage as checking TPM gave: tpm says no\n")
-}
-
-func (s *deviceMgrInstallModeSuite) TestInstallCheckEncryptedErrorsLogsHook(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	logbuf, restore := logger.MockLogger()
-	defer restore()
-
-	mockModel := s.makeModelAssertionInState(c, "my-brand", "my-model", checkEncryptionModelHeaders)
-	s.makeMockInstalledPcGadget(c, "", "")
-	// mock kernel installed but no hook or handle so checkEncryption
-	// will fail
-	makeInstalledMockKernelSnap(c, s.state, kernelYamlWithFdeSetup)
-
-	deviceCtx := &snapstatetest.TrivialDeviceContext{DeviceModel: mockModel}
-	_, err := devicestate.DeviceManagerCheckEncryption(s.mgr, s.state, deviceCtx, secboot.TPMProvisionFull)
-	c.Check(err, IsNil)
-	c.Check(logbuf.String(), Matches, "(?s).*: not encrypting device storage as querying kernel fde-setup hook did not succeed:.*\n")
 }
 
 func (s *deviceMgrInstallModeSuite) TestInstallHappyLogfiles(c *C) {
@@ -2408,7 +2176,7 @@ func (s *deviceMgrInstallModeSuite) doRunFactoryResetChange(c *C, model *asserts
 	})
 	defer restore()
 
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error {
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error {
 		if tc.tpm {
 			return nil
 		} else {
@@ -3017,7 +2785,7 @@ func (s *deviceMgrInstallModeSuite) TestFactoryResetExpectedTasks(c *C) {
 	restore := release.MockOnClassic(false)
 	defer restore()
 
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
 		c.Check(tpmMode, Equals, secboot.TPMPartialReprovision)
 		return fmt.Errorf("TPM not available")
 	})
@@ -3078,7 +2846,7 @@ func (s *deviceMgrInstallModeSuite) TestFactoryResetInstallDeviceHook(c *C) {
 	restore := release.MockOnClassic(false)
 	defer restore()
 
-	restore = devicestate.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
+	restore = installLogic.MockSecbootCheckTPMKeySealingSupported(func(tpmMode secboot.TPMProvisionMode) error {
 		c.Check(tpmMode, Equals, secboot.TPMPartialReprovision)
 		return fmt.Errorf("TPM not available")
 	})
@@ -3233,432 +3001,6 @@ func (s *deviceMgrInstallModeSuite) TestFactoryResetWritesTimesyncdClock(c *C) {
 	c.Assert(err, IsNil)
 	c.Check(fi.ModTime().Round(time.Second), Equals, now.Round(time.Second))
 	c.Check(fi.Size(), Equals, int64(0))
-}
-
-func (s *deviceMgrInstallModeSuite) TestEncryptionSupportInfoForceUnencrypted(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	kernelInfo := makeInstalledMockKernelSnap(c, s.state, kernelYamlNoFdeSetup)
-	gadgetSnapInfo := s.makeMockInstalledPcGadget(c, "", "")
-	gadgetInfo, err := gadget.ReadInfo(gadgetSnapInfo.MountDir(), nil)
-	c.Assert(err, IsNil)
-
-	var testCases = []struct {
-		grade, storageSafety, forceUnencrypted string
-		tpmErr                                 error
-
-		expected devicestate.EncryptionSupportInfo
-	}{
-		{
-			"dangerous", "", "", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		},
-		{
-			"dangerous", "", "force-unencrypted", nil,
-			devicestate.EncryptionSupportInfo{
-				// Encryption is forcefully disabled
-				// here so no further
-				// availability/enc-type checks are
-				// performed.
-				Available: false, Disabled: true,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeNone,
-			},
-		},
-		{
-			"dangerous", "", "force-unencrypted", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				// Encryption is forcefully disabled
-				// here so the "no tpm" error is never visible
-				Available: false, Disabled: true,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeNone,
-			},
-		},
-		// not possible to disable encryption on non-dangerous devices
-		{
-			"signed", "", "", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		},
-		{
-			"signed", "", "force-unencrypted", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		},
-		{
-			"signed", "", "force-unencrypted", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:      asserts.StorageSafetyPreferEncrypted,
-				Type:               secboot.EncryptionTypeNone,
-				UnavailableWarning: "not encrypting device storage as checking TPM gave: no tpm",
-			},
-		},
-		{
-			"secured", "", "", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		},
-		{
-			"secured", "", "force-unencrypted", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		},
-		{
-			"secured", "", "force-unencrypted", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:  asserts.StorageSafetyEncrypted,
-				Type:           secboot.EncryptionTypeNone,
-				UnavailableErr: fmt.Errorf("cannot encrypt device storage as mandated by model grade secured: no tpm"),
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		restore := devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return tc.tpmErr })
-		defer restore()
-
-		mockModel := s.makeModelAssertionInState(c, "my-brand", "my-model", map[string]interface{}{
-			"display-name":   "my model",
-			"architecture":   "amd64",
-			"base":           "core20",
-			"grade":          tc.grade,
-			"storage-safety": tc.storageSafety,
-			"snaps": []interface{}{
-				map[string]interface{}{
-					"name":            "pc-kernel",
-					"id":              pcKernelSnapID,
-					"type":            "kernel",
-					"default-channel": "20",
-				},
-				map[string]interface{}{
-					"name":            "pc",
-					"id":              pcSnapID,
-					"type":            "gadget",
-					"default-channel": "20",
-				}},
-		})
-		forceUnencryptedPath := filepath.Join(boot.InitramfsUbuntuSeedDir, ".force-unencrypted")
-		if tc.forceUnencrypted == "" {
-			os.Remove(forceUnencryptedPath)
-		} else {
-			err := os.MkdirAll(filepath.Dir(forceUnencryptedPath), 0755)
-			c.Assert(err, IsNil)
-			err = ioutil.WriteFile(forceUnencryptedPath, nil, 0644)
-			c.Assert(err, IsNil)
-		}
-
-		res, err := devicestate.DeviceManagerEncryptionSupportInfo(s.mgr, mockModel, secboot.TPMProvisionFull, kernelInfo, gadgetInfo)
-		c.Assert(err, IsNil)
-		c.Check(res, DeepEquals, tc.expected, Commentf("%v", tc))
-	}
-}
-
-var gadgetWithoutUbuntuSave = &gadget.Info{
-	Volumes: map[string]*gadget.Volume{
-		"pc": {
-			Name:       "pc",
-			Schema:     "mbr",
-			Bootloader: "grub",
-			Structure: []gadget.VolumeStructure{
-				{VolumeName: "ubuntu-seed", Name: "ubuntu-seed", Label: "ubuntu-seed", Size: 700 * quantity.SizeMiB, Role: "system-seed", Filesystem: "vfat"},
-				{VolumeName: "ubuntu-data", Name: "ubuntu-data", Label: "ubuntu-data", Size: 700 * quantity.SizeMiB, Role: "system-data", Filesystem: "ext4"},
-			},
-		},
-	},
-}
-
-var gadgetUC20 = &gadget.Info{
-	Volumes: map[string]*gadget.Volume{
-		"pc": {
-			Name:       "pc",
-			Schema:     "mbr",
-			Bootloader: "grub",
-			Structure: []gadget.VolumeStructure{
-				{VolumeName: "ubuntu-seed", Name: "ubuntu-seed", Label: "ubuntu-seed", Size: 700 * quantity.SizeMiB, Role: "system-seed", Filesystem: "vfat"},
-				{VolumeName: "ubuntu-data", Name: "ubuntu-data", Label: "ubuntu-data", Size: 700 * quantity.SizeMiB, Role: "system-data", Filesystem: "ext4"},
-				{VolumeName: "ubuntu-save", Name: "ubuntu-save", Label: "ubuntu-save", Size: 5 * quantity.SizeMiB, Role: "system-save", Filesystem: "ext4"},
-			},
-		},
-	},
-}
-
-func (s *deviceMgrInstallModeSuite) TestEncryptionSupportInfoGadgetIncompatibleWithEncryption(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	restore := devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return nil })
-	defer restore()
-
-	kernelInfo := makeInstalledMockKernelSnap(c, s.state, kernelYamlNoFdeSetup)
-	var testCases = []struct {
-		grade, storageSafety string
-		gadgetInfo           *gadget.Info
-
-		expected devicestate.EncryptionSupportInfo
-	}{
-		{
-			"dangerous", "", gadgetUC20,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"dangerous", "", gadgetWithoutUbuntuSave,
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:      asserts.StorageSafetyPreferEncrypted,
-				Type:               secboot.EncryptionTypeNone,
-				UnavailableWarning: "cannot use encryption with the gadget, disabling encryption: gadget does not support encrypted data: required partition with system-save role is missing",
-			},
-		}, {
-			"dangerous", "encrypted", gadgetWithoutUbuntuSave,
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:  asserts.StorageSafetyEncrypted,
-				Type:           secboot.EncryptionTypeNone,
-				UnavailableErr: fmt.Errorf("cannot use encryption with the gadget: gadget does not support encrypted data: required partition with system-save role is missing"),
-			},
-		}, {
-			"signed", "", gadgetUC20,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"signed", "", gadgetWithoutUbuntuSave,
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:      asserts.StorageSafetyPreferEncrypted,
-				Type:               secboot.EncryptionTypeNone,
-				UnavailableWarning: "cannot use encryption with the gadget, disabling encryption: gadget does not support encrypted data: required partition with system-save role is missing",
-			},
-		}, {
-			"signed", "encrypted", gadgetWithoutUbuntuSave,
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:  asserts.StorageSafetyEncrypted,
-				Type:           secboot.EncryptionTypeNone,
-				UnavailableErr: fmt.Errorf("cannot use encryption with the gadget: gadget does not support encrypted data: required partition with system-save role is missing"),
-			},
-		}, {
-			"secured", "", gadgetUC20,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"secured", "", gadgetWithoutUbuntuSave,
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:  asserts.StorageSafetyEncrypted,
-				Type:           secboot.EncryptionTypeNone,
-				UnavailableErr: fmt.Errorf("cannot use encryption with the gadget: gadget does not support encrypted data: required partition with system-save role is missing"),
-			},
-		},
-	}
-	for _, tc := range testCases {
-		mockModel := s.makeModelAssertionInState(c, "my-brand", "my-model", map[string]interface{}{
-			"display-name":   "my model",
-			"architecture":   "amd64",
-			"base":           "core20",
-			"grade":          tc.grade,
-			"storage-safety": tc.storageSafety,
-			"snaps": []interface{}{
-				map[string]interface{}{
-					"name":            "pc-kernel",
-					"id":              pcKernelSnapID,
-					"type":            "kernel",
-					"default-channel": "20",
-				},
-				map[string]interface{}{
-					"name":            "pc",
-					"id":              pcSnapID,
-					"type":            "gadget",
-					"default-channel": "20",
-				}},
-		})
-		res, err := devicestate.DeviceManagerEncryptionSupportInfo(s.mgr, mockModel, secboot.TPMProvisionFull, kernelInfo, tc.gadgetInfo)
-		c.Assert(err, IsNil)
-		c.Check(res, DeepEquals, tc.expected, Commentf("%v", tc))
-	}
-}
-
-func (s *deviceMgrInstallModeSuite) TestEncryptionSupportInfoWithTPM(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	kernelInfo := makeInstalledMockKernelSnap(c, s.state, kernelYamlNoFdeSetup)
-	gadgetSnapInfo := s.makeMockInstalledPcGadget(c, "", "")
-	gadgetInfo, err := gadget.ReadInfo(gadgetSnapInfo.MountDir(), nil)
-	c.Assert(err, IsNil)
-
-	var testCases = []struct {
-		grade, storageSafety string
-		tpmErr               error
-
-		expected devicestate.EncryptionSupportInfo
-	}{
-		{
-			"dangerous", "", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"dangerous", "", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:      asserts.StorageSafetyPreferEncrypted,
-				Type:               secboot.EncryptionTypeNone,
-				UnavailableWarning: "not encrypting device storage as checking TPM gave: no tpm",
-			},
-		}, {
-			"dangerous", "encrypted", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"dangerous", "encrypted", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:  asserts.StorageSafetyEncrypted,
-				Type:           secboot.EncryptionTypeNone,
-				UnavailableErr: fmt.Errorf("cannot encrypt device storage as mandated by encrypted storage-safety model option: no tpm"),
-			},
-		},
-		{
-			"dangerous", "prefer-unencrypted", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferUnencrypted,
-				// Note that encryption type is set to what is available
-				Type: secboot.EncryptionTypeLUKS,
-			},
-		},
-		{
-			"signed", "", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"signed", "", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:      asserts.StorageSafetyPreferEncrypted,
-				Type:               secboot.EncryptionTypeNone,
-				UnavailableWarning: "not encrypting device storage as checking TPM gave: no tpm",
-			},
-		}, {
-			"signed", "encrypted", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"signed", "prefer-unencrypted", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyPreferUnencrypted,
-				// Note that encryption type is set to what is available
-				Type: secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"signed", "encrypted", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:  asserts.StorageSafetyEncrypted,
-				Type:           secboot.EncryptionTypeNone,
-				UnavailableErr: fmt.Errorf("cannot encrypt device storage as mandated by encrypted storage-safety model option: no tpm"),
-			},
-		}, {
-			"secured", "encrypted", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"secured", "encrypted", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:  asserts.StorageSafetyEncrypted,
-				Type:           secboot.EncryptionTypeNone,
-				UnavailableErr: fmt.Errorf("cannot encrypt device storage as mandated by model grade secured: no tpm"),
-			},
-		}, {
-			"secured", "", nil,
-			devicestate.EncryptionSupportInfo{
-				Available: true, Disabled: false,
-				StorageSafety: asserts.StorageSafetyEncrypted,
-				Type:          secboot.EncryptionTypeLUKS,
-			},
-		}, {
-			"secured", "", fmt.Errorf("no tpm"),
-			devicestate.EncryptionSupportInfo{
-				Available: false, Disabled: false,
-				StorageSafety:  asserts.StorageSafetyEncrypted,
-				Type:           secboot.EncryptionTypeNone,
-				UnavailableErr: fmt.Errorf("cannot encrypt device storage as mandated by model grade secured: no tpm"),
-			},
-		},
-	}
-	for _, tc := range testCases {
-		restore := devicestate.MockSecbootCheckTPMKeySealingSupported(func(secboot.TPMProvisionMode) error { return tc.tpmErr })
-		defer restore()
-
-		mockModel := s.makeModelAssertionInState(c, "my-brand", "my-model", map[string]interface{}{
-			"display-name":   "my model",
-			"architecture":   "amd64",
-			"base":           "core20",
-			"grade":          tc.grade,
-			"storage-safety": tc.storageSafety,
-			"snaps": []interface{}{
-				map[string]interface{}{
-					"name":            "pc-kernel",
-					"id":              pcKernelSnapID,
-					"type":            "kernel",
-					"default-channel": "20",
-				},
-				map[string]interface{}{
-					"name":            "pc",
-					"id":              pcSnapID,
-					"type":            "gadget",
-					"default-channel": "20",
-				}},
-		})
-		res, err := devicestate.DeviceManagerEncryptionSupportInfo(s.mgr, mockModel, secboot.TPMProvisionFull, kernelInfo, gadgetInfo)
-		c.Assert(err, IsNil)
-		c.Check(res, DeepEquals, tc.expected, Commentf("%v", tc))
-	}
 }
 
 func (s *deviceMgrInstallModeSuite) TestInstallWithUbuntuSaveSnapFoldersHappy(c *C) {
